fix: 라우터 문제 수정
TanStack/router#425 문제 해결
